# docs/openapi.yaml
openapi: 3.1.0
info:
  title: 감정지도 MVP API 계약 초안
  version: 0.2.0-draft
  description: '기준: MVP 1.4-final + ERD 1.0 + FE v2 + 사용자 이메일·비밀번호 전용 로그인 확정. L03 이미지 업로드·수명주기 primitive와 해당 HTTP 경로만 현재 구현 범위이며, 나머지 신규 DTO·저장구조·토큰 TTL·수치 한도는 검토용 계약 초안이다. 예시는 mock이며 서비스 실데이터/설정이 아니다. 참여자용 22개 경로에 /v1 접두어를 적용한다.'
servers:
- url: /
  description: 배포 origin을 설정. paths에 /v1이 포함되어 있으므로 서버 URL에 중복 추가하지 않음
...
security:
- bearerAuth: []
paths:
...
  /v1/memories/analyze:
    post:
      operationId: analyzeMemory
      summary: 본문의 분위기·카테고리 AI 제안
      tags:
      - Memories
      description: 본문만 분석하며 계정 취향으로 빈 축을 채우지 않는다. 유효한 요청의 AI 분석 실패는200+FAILED+null축으로 수동 입력을 지원한다. 서버 자체 장애503와
        구분. analysisToken은 서버 서명 확인값 제안이며 최종 저장에서 위변조·본문·사용자·만료 검사. 수동 저장은 토큰null로 가능하다.
      x-contract-status: PROPOSED
      x-onboarding-required: true
      responses:
        '200':
          description: 성공
          headers:
            Cache-Control:
              schema:
                type: string
              description: 개인 응답은 private, no-store
            X-Request-Id:
              schema:
                type: string
              description: 서버 추적 식별자
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalyzeResponse'
              examples:
                complete:
                  value:
                    atmospheres:
                      CROWD_LEVEL: -1
                      SPATIAL_FEEL: -1
                      COMPANY_FIT: 1
                      STAY_STYLE: -1
                    categories:
                    - CAFE
                    - STUDY_WORK
                    categoryStatus: CLASSIFIED
                    atmosphereStatus: SUCCEEDED
                    analysisToken: mock-signed-analysis-receipt-not-a-live-token
                    expiresAt: '2026-09-19T11:15:00+09:00'
                    warnings: []
                partial:
                  value:
                    atmospheres:
                      CROWD_LEVEL: -1
                      SPATIAL_FEEL: -1
                      COMPANY_FIT: 1
                      STAY_STYLE: null
                    categories:
                    - CAFE
                    - STUDY_WORK
                    categoryStatus: CLASSIFIED
                    atmosphereStatus: PARTIAL
                    analysisToken: mock-signed-analysis-receipt-not-a-live-token
                    expiresAt: '2026-09-19T11:15:00+09:00'
                    warnings:
                    - ATMOSPHERE_NEEDS_INPUT
                upstreamFailure:
                  value:
                    atmospheres:
                      CROWD_LEVEL: null
                      SPATIAL_FEEL: null
                      COMPANY_FIT: null
                      STAY_STYLE: null
                    categories: []
                    categoryStatus: FAILED
                    atmosphereStatus: FAILED
                    analysisToken: mock-signed-analysis-receipt-not-a-live-token
                    expiresAt: '2026-09-19T11:15:00+09:00'
                    warnings:
                    - ANALYSIS_FAILED
        '401':
          description: AUTH_REQUIRED, TOKEN_EXPIRED, INVALID_TOKEN
          headers:
            Cache-Control:
              schema:
                type: string
              description: 개인 응답은 private, no-store
            X-Request-Id:
              schema:
                type: string
              description: 서버 추적 식별자
            WWW-Authenticate:
              schema:
                type: string
              description: Bearer 인증 실패 응답 challenge
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              examples:
                AUTH_REQUIRED:
                  value:
                    code: AUTH_REQUIRED
                    message: 로그인이 필요합니다
                    retryAfterSeconds: null
                    fieldErrors: []
                    requestId: 80000000-0000-4000-8000-000000000001
                TOKEN_EXPIRED:
                  value:
                    code: TOKEN_EXPIRED
                    message: 로그인이 만료되었습니다. 다시 로그인해 주세요
                    retryAfterSeconds: null
                    fieldErrors: []
                    requestId: 80000000-0000-4000-8000-000000000001
                INVALID_TOKEN:
                  value:
                    code: INVALID_TOKEN
                    message: 인증 정보를 확인할 수 없습니다
                    retryAfterSeconds: null
                    fieldErrors: []
                    requestId: 80000000-0000-4000-8000-000000000001
        '403':
          description: INVITATION_REQUIRED, ACCOUNT_SUSPENDED, ACCOUNT_CLOSED, ONBOARDING_REQUIRED
          headers:
            Cache-Control:
              schema:
                type: string
              description: 개인 응답은 private, no-store
            X-Request-Id:
              schema:
                type: string
              description: 서버 추적 식별자
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              examples:
                INVITATION_REQUIRED:
                  value:
                    code: INVITATION_REQUIRED
                    message: 초대된 계정만 사용할 수 있습니다
                    retryAfterSeconds: null
                    fieldErrors: []
                    requestId: 80000000-0000-4000-8000-000000000001
                ACCOUNT_SUSPENDED:
                  value:
                    code: ACCOUNT_SUSPENDED
                    message: 이용이 제한된 계정입니다
                    retryAfterSeconds: null
                    fieldErrors: []
                    requestId: 80000000-0000-4000-8000-000000000001
                ACCOUNT_CLOSED:
                  value:
                    code: ACCOUNT_CLOSED
                    message: 종료된 계정입니다
                    retryAfterSeconds: null
                    fieldErrors: []
                    requestId: 80000000-0000-4000-8000-000000000001
                ONBOARDING_REQUIRED:
                  value:
                    code: ONBOARDING_REQUIRED
                    message: 초기 설정을 완료해 주세요
                    retryAfterSeconds: null
                    fieldErrors: []
                    requestId: 80000000-0000-4000-8000-000000000001
        '503':
          description: SERVICE_UNAVAILABLE
          headers:
            Cache-Control:
              schema:
                type: string
              description: 개인 응답은 private, no-store
            X-Request-Id:
              schema:
                type: string
              description: 서버 추적 식별자
            Retry-After:
              schema:
                type: string
                pattern: ^[0-9]+$
              description: 재시도 가능한 경우에만 남은 초. body retryAfterSeconds와 일치
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              examples:
                SERVICE_UNAVAILABLE:
                  value:
                    code: SERVICE_UNAVAILABLE
                    message: 일시적으로 요청을 처리할 수 없습니다
                    retryAfterSeconds: null
                    fieldErrors: []
                    requestId: 80000000-0000-4000-8000-000000000001
        '400':
          description: INVALID_JSON, DUPLICATE_JSON_KEY, INVALID_REQUEST
          headers:
            Cache-Control:
              schema:
                type: string
              description: 개인 응답은 private, no-store
            X-Request-Id:
              schema:
                type: string
              description: 서버 추적 식별자
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              examples:
                INVALID_JSON:
                  value:
                    code: INVALID_JSON
                    message: JSON 형식이 올바르지 않습니다
                    retryAfterSeconds: null
                    fieldErrors: []
                    requestId: 80000000-0000-4000-8000-000000000001
                DUPLICATE_JSON_KEY:
                  value:
                    code: DUPLICATE_JSON_KEY
                    message: 중복된 JSON 키가 있습니다
                    retryAfterSeconds: null
                    fieldErrors: []
                    requestId: 80000000-0000-4000-8000-000000000001
                INVALID_REQUEST:
                  value:
                    code: INVALID_REQUEST
                    message: 요청 형식이 올바르지 않습니다
                    retryAfterSeconds: null
                    fieldErrors: []
                    requestId: 80000000-0000-4000-8000-000000000001
        '422':
          description: VALIDATION_ERROR
          headers:
            Cache-Control:
              schema:
                type: string
              description: 개인 응답은 private, no-store
            X-Request-Id:
              schema:
                type: string
              description: 서버 추적 식별자
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              examples:
                VALIDATION_ERROR:
                  value:
                    code: VALIDATION_ERROR
                    message: 입력값을 확인해 주세요
                    retryAfterSeconds: null
                    fieldErrors: []
                    requestId: 80000000-0000-4000-8000-000000000001
        '429':
          description: RATE_LIMITED
          headers:
            Cache-Control:
              schema:
                type: string
              description: 개인 응답은 private, no-store
            X-Request-Id:
              schema:
                type: string
              description: 서버 추적 식별자
            Retry-After:
              schema:
                type: string
                pattern: ^[0-9]+$
              description: 재시도 가능한 경우에만 남은 초. body retryAfterSeconds와 일치
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              examples:
                RATE_LIMITED:
                  value:
                    code: RATE_LIMITED
                    message: 요청이 너무 많습니다. 잠시 후 다시 시도해 주세요
                    retryAfterSeconds: null
                    fieldErrors: []
                    requestId: 80000000-0000-4000-8000-000000000001
        default:
          description: 예상하지 못한 오류도 동일 ApiError 형식. 실제 성공/실패는 HTTP 상태를 따른다.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnalyzeRequest'
            examples:
              valid:
                value:
                  content: 작은 카페의 조용한 창가에서 친구와 함께 책을 읽었다. 작업용 책상이 있고 오래 머물기 편한 아늑한 공간이었다.
...
  /v1/memories/{id}/image:
    get:
      operationId: getMemoryImage
      summary: 권한 확인 후 경험 이미지
      tags:
      - Images
      description: Bearer 헤더 필수. 존재하지 않음/권한 없음/사진 없음404. 과거 수신 또는 소유가 확인된 삭제·숨김은410. 파일 메타데이터는 있지만 IO 장애면503. 공개
        리다이렉트·토큰 query 없음. Cache-Control:private,no-store; 고정 안전 파일명; X-Content-Type-Options:nosniff.
      x-contract-status: PROPOSED
      x-onboarding-required: true
      responses:
        '200':
          description: 성공
          headers:
            Cache-Control:
              schema:
                type: string
              description: 개인 응답은 private, no-store
            X-Request-Id:
              schema:
                type: string
              description: 서버 추적 식별자
            Content-Disposition:
              schema:
                type: string
              description: inline; filename="experience.jpg" (또는png), 원본 파일명·계정 식별정보 없음
            X-Content-Type-Options:
              schema:
                type: string
                const: nosniff
          content:
            image/jpeg: {}
            image/png: {}
        '401':
          description: AUTH_REQUIRED, TOKEN_EXPIRED, INVALID_TOKEN
          headers:
            Cache-Control:
              schema:
                type: string
              description: 개인 응답은 private, no-store
            X-Request-Id:
              schema:
                type: string
              description: 서버 추적 식별자
            WWW-Authenticate:
              schema:
                type: string
              description: Bearer 인증 실패 응답 challenge
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              examples:
                AUTH_REQUIRED:
                  value:
                    code: AUTH_REQUIRED
                    message: 로그인이 필요합니다
                    retryAfterSeconds: null
                    fieldErrors: []
                    requestId: 80000000-0000-4000-8000-000000000001
                TOKEN_EXPIRED:
                  value:
                    code: TOKEN_EXPIRED
                    message: 로그인이 만료되었습니다. 다시 로그인해 주세요
                    retryAfterSeconds: null
                    fieldErrors: []
                    requestId: 80000000-0000-4000-8000-000000000001
                INVALID_TOKEN:
                  value:
                    code: INVALID_TOKEN
                    message: 인증 정보를 확인할 수 없습니다
                    retryAfterSeconds: null
                    fieldErrors: []
                    requestId: 80000000-0000-4000-8000-000000000001
        '403':
          description: INVITATION_REQUIRED, ACCOUNT_SUSPENDED, ACCOUNT_CLOSED, ONBOARDING_REQUIRED
          headers:
            Cache-Control:
              schema:
                type: string
              description: 개인 응답은 private, no-store
            X-Request-Id:
              schema:
                type: string
              description: 서버 추적 식별자
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              examples:
                INVITATION_REQUIRED:
                  value:
                    code: INVITATION_REQUIRED
                    message: 초대된 계정만 사용할 수 있습니다
                    retryAfterSeconds: null
                    fieldErrors: []
                    requestId: 80000000-0000-4000-8000-000000000001
                ACCOUNT_SUSPENDED:
                  value:
                    code: ACCOUNT_SUSPENDED
                    message: 이용이 제한된 계정입니다
                    retryAfterSeconds: null
                    fieldErrors: []
                    requestId: 80000000-0000-4000-8000-000000000001
                ACCOUNT_CLOSED:
                  value:
                    code: ACCOUNT_CLOSED
                    message: 종료된 계정입니다
                    retryAfterSeconds: null
                    fieldErrors: []
                    requestId: 80000000-0000-4000-8000-000000000001
                ONBOARDING_REQUIRED:
                  value:
                    code: ONBOARDING_REQUIRED
                    message: 초기 설정을 완료해 주세요
                    retryAfterSeconds: null
                    fieldErrors: []
                    requestId: 80000000-0000-4000-8000-000000000001
        '503':
          description: SERVICE_UNAVAILABLE, IMAGE_FILE_UNAVAILABLE
          headers:
            Cache-Control:
              schema:
                type: string
              description: 개인 응답은 private, no-store
            X-Request-Id:
              schema:
                type: string
              description: 서버 추적 식별자
            Retry-After:
              schema:
                type: string
                pattern: ^[0-9]+$
              description: 재시도 가능한 경우에만 남은 초. body retryAfterSeconds와 일치
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              examples:
                SERVICE_UNAVAILABLE:
                  value:
                    code: SERVICE_UNAVAILABLE
                    message: 일시적으로 요청을 처리할 수 없습니다
                    retryAfterSeconds: null
                    fieldErrors: []
                    requestId: 80000000-0000-4000-8000-000000000001
                IMAGE_FILE_UNAVAILABLE:
                  value:
                    code: IMAGE_FILE_UNAVAILABLE
                    message: 이미지 파일을 불러오지 못했습니다
                    retryAfterSeconds: null
                    fieldErrors: []
                    requestId: 80000000-0000-4000-8000-000000000001
        '404':
          description: RESOURCE_NOT_FOUND
          headers:
            Cache-Control:
              schema:
                type: string
              description: 개인 응답은 private, no-store
            X-Request-Id:
              schema:
                type: string
              description: 서버 추적 식별자
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              examples:
                RESOURCE_NOT_FOUND:
                  value:
                    code: RESOURCE_NOT_FOUND
                    message: 요청한 항목을 찾을 수 없습니다
                    retryAfterSeconds: null
                    fieldErrors: []
                    requestId: 80000000-0000-4000-8000-000000000001
        '410':
          description: MEMORY_UNAVAILABLE
          headers:
            Cache-Control:
              schema:
                type: string
              description: 개인 응답은 private, no-store
            X-Request-Id:
              schema:
                type: string
              description: 서버 추적 식별자
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              examples:
                MEMORY_UNAVAILABLE:
                  value:
                    code: MEMORY_UNAVAILABLE
                    message: 더 이상 열람할 수 없는 경험입니다
                    retryAfterSeconds: null
                    fieldErrors: []
                    requestId: 80000000-0000-4000-8000-000000000001
        default:
          description: 예상하지 못한 오류도 동일 ApiError 형식. 실제 성공/실패는 HTTP 상태를 따른다.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
      parameters:
      - name: id
        in: path
        required: true
        description: 경험ID
        schema:
          type: string
          format: uuid
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: 서비스 access token. JWT인지 여부는 클라이언트가 해석하지 않는다.
  schemas:
...
    PlaceCategoryCode:
      type: string
      enum:
      - CAFE
      - RESTAURANT
      - BAR
      - PARK_WALK
      - CULTURE
      - STUDY_WORK
      - SHOPPING
      - OTHER
...
    AnalyzedAtmospheres:
      type: object
      properties:
        CROWD_LEVEL:
          anyOf:
          - type: integer
            enum:
            - -1
            - 1
          - type: 'null'
        SPATIAL_FEEL:
          anyOf:
          - type: integer
            enum:
            - -1
            - 1
          - type: 'null'
        COMPANY_FIT:
          anyOf:
          - type: integer
            enum:
            - -1
            - 1
          - type: 'null'
        STAY_STYLE:
          anyOf:
          - type: integer
            enum:
            - -1
            - 1
          - type: 'null'
      required:
      - CROWD_LEVEL
      - SPATIAL_FEEL
      - COMPANY_FIT
      - STAY_STYLE
      additionalProperties: false
      description: AI 분석 중에만 null 허용. 최종 저장에는 null 불가.
    CategoryCodes:
      type: array
      items:
        $ref: '#/components/schemas/PlaceCategoryCode'
      maxItems: 3
      uniqueItems: true
...
    FieldError:
      type: object
      properties:
        field:
          type: string
        reason:
          type: string
          enum:
          - REQUIRED
          - INVALID_VALUE
          - DUPLICATE
          - OUT_OF_RANGE
          - UNKNOWN_FIELD
          - TOO_LONG
          - IMMUTABLE
      required:
      - field
      - reason
      additionalProperties: false
    ErrorCode:
      type: string
      enum:
      - INVALID_CREDENTIALS
      - EMAIL_ALREADY_EXISTS
      - TOO_MANY_ATTEMPTS
      - AUTH_REQUIRED
      - TOKEN_EXPIRED
      - INVALID_TOKEN
      - INVITATION_REQUIRED
      - ACCOUNT_SUSPENDED
      - ACCOUNT_CLOSED
      - ONBOARDING_REQUIRED
      - INVALID_JSON
      - DUPLICATE_JSON_KEY
      - INVALID_REQUEST
      - VALIDATION_ERROR
      - INVALID_ATMOSPHERES
      - INVALID_CATEGORIES
      - IMMUTABLE_FIELD
      - ONBOARDING_ALREADY_COMPLETED
      - PREFERENCE_VERSION_CONFLICT
      - RESOURCE_NOT_FOUND
      - MEMORY_UNAVAILABLE
      - IMAGE_NOT_FOUND
      - IMAGE_UPLOAD_EXPIRED
      - IMAGE_ALREADY_ATTACHED
      - IMAGE_TOO_LARGE
      - UNSUPPORTED_IMAGE_TYPE
      - INVALID_IMAGE
      - IMAGE_DIMENSIONS_EXCEEDED
      - IMAGE_STORAGE_UNAVAILABLE
      - IMAGE_FILE_UNAVAILABLE
      - ANALYSIS_TOKEN_INVALID
      - ANALYSIS_TOKEN_EXPIRED
      - ANALYSIS_CONTENT_MISMATCH
      - DAILY_WRITE_LIMIT_EXCEEDED
      - RATE_LIMITED
      - INVALID_CURSOR
      - CURSOR_CONTEXT_MISMATCH
      - INVALID_BBOX
      - PLACE_COORDINATE_MISMATCH
      - IDEMPOTENCY_KEY_REQUIRED
      - IDEMPOTENCY_KEY_REUSED
      - REQUEST_IN_PROGRESS
      - COPY_FAILED
      - CONFIGURATION_UNAVAILABLE
      - SERVICE_UNAVAILABLE
      - SERVER_ERROR
    ApiError:
      type: object
      properties:
        code:
          $ref: '#/components/schemas/ErrorCode'
        message:
          type: string
        retryAfterSeconds:
          anyOf:
          - type: integer
            minimum: 0
          - type: 'null'
        fieldErrors:
          type: array
          items:
            $ref: '#/components/schemas/FieldError'
        requestId:
          type: string
          minLength: 1
      required:
      - code
      - message
      - retryAfterSeconds
      - fieldErrors
      - requestId
      additionalProperties: false
      description: code로 분기. retryAfterSeconds는 없으면 null이며 429 응답의 Retry-After 헤더와 동일하다. message는 사용자 표시용으로 SQL/경로/작성자
        정보 금지.
...
    AnalyzeRequest:
      type: object
      properties:
        content:
          type: string
          minLength: 1
          pattern: \S
      required:
      - content
      additionalProperties: false
    AnalyzeResponse:
      type: object
      properties:
        atmospheres:
          $ref: '#/components/schemas/AnalyzedAtmospheres'
        categories:
          $ref: '#/components/schemas/CategoryCodes'
        categoryStatus:
          type: string
          enum:
          - CLASSIFIED
          - UNCLASSIFIED
          - FAILED
        atmosphereStatus:
          type: string
          enum:
          - SUCCEEDED
          - PARTIAL
          - FAILED
        analysisToken:
          type: string
          minLength: 1
        expiresAt:
          type: string
          format: date-time
        warnings:
          type: array
          items:
            type: string
            enum:
            - ATMOSPHERE_NEEDS_INPUT
            - CATEGORY_UNCLASSIFIED
            - ANALYSIS_FAILED
      required:
      - atmospheres
      - categories
      - categoryStatus
      - atmosphereStatus
      - analysisToken
      - expiresAt
      - warnings
      additionalProperties: false
      description: analysisToken은 서버 서명 확인값 제안. 본문 해시·사용자·AI 제안/실행상태·모델 버전·만료를 검증한다. 실패한 AI 호출도 수동 입력을 위한 200+FAILED로
        표현 가능. 최종 축 null은 불가.
...